+++ iota-client/Cargo.toml
@@ -19,3 +19,5 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 anyhow = "1.0.26"
 tokio = { version = "0.2.9", features = ["macros"] }
+bee-bundle = { git = "https://github.com/wusyong/bee-p", branch = "yuwei" }
+bee-ternary = { git = "https://github.com/wusyong/bee-p", branch = "yuwei" }
